@@ -14,7 +14,6 @@ classifiers = [
   "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
   "Operating System :: MacOS :: MacOS X",
   "Operating System :: POSIX :: Linux",
-  "Programming Language :: Python :: 3.7",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
@@ -36,7 +35,7 @@ multizone = "multizone:main"
 
 [tool.black]
 line-length = 108
-target-version = ['py37']
+target-version = ['py38']
 
 [tool.coverage.run]
 include = ["multizone.py"]
@@ -49,10 +48,14 @@ disable = "R0912"
 
 [tool.ruff]
 fixable = ["D", "I", "COM"]
-ignore = ["S101"]
+ignore = [
+  "S101",  # assert
+  "D213",  # multi-line-summary-second-line incompatible with multi-line-summary-first-line
+  "D203"  # one-blank-line-before-class" incompatible with no-blank-line-before-class
+]
 line-length = 108
 select = ["ALL"]
-target-version = "py37"
+target-version = "py38"
 
 [tool.ruff.per-file-ignores]
 "test_mz.py" = ["PLR2004"]
